+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Add command line option to save file with a command list
 - Create command list
 - Add help command
+- Add explore command